epiccodereview: fix 006-ladder sweep1 (30 issues)
Tests:
- cross_encoder.rs: add #![cfg(jxl-codec)] guard (CE-01)
- cross_encoder.rs: compare sRGB-only, tighten tolerance 0.35→0.03 (L001/CE-02/SEC-007/CTR-002)
- cross_encoder.rs: ct6 vacuous test → #[ignore] with minimal panic-safety body (L003/CE-03/SEC-006/CTR-003)
- cross_encoder.rs: aspect-ratio tolerance 0.1→0.02 (L007)
- jxl_lowlevel_progressive.rs: any()→all() for intermediate-frame check (PF-01)
- jxl_lowlevel_progressive.rs: remove dead redundant assert inside if-saw_non_final (L002/PF-02)

Security:
- fast-jpeg/lib.rs: checked_mul for L8 capacity — avoids 32-bit wrap (SEC-001)
- fast-jpeg/lib.rs: 400MP decompression-bomb guard in decode_scaled (SEC-003)
- jxl_casadecoder.rs: cast u32→usize before multiply in JXTC tile copy (SEC-004)

Concurrency:
- pipeline.rs: PERCEPTUAL_GRID init with borrow_mut once, hot loop with borrow (CONC-03)
- ljpeg.rs: Rc<HuffTable>→Arc<HuffTable> in DHT_CACHE thread-local (CONC-04)

API/Docs:
- fast-jpeg/lib.rs: add take_data(self) consuming accessor (FJ-01/CTR-005/PERF-001)
- fast-jpeg/lib.rs: UnsupportedFormat: prefix on CMYK/L16 errors (FJ-02)
- pipeline.rs: clarify process_into docstring (elements not bytes) (CTR-004)

Benchmarks:
- traversal_fusion_flipflop.rs: closures use injected arg (TF-01); chunks→chunks_exact (PERF-005); interleave A/B with start-rotation, report median (L004)
- perceptual_flipflop.rs: alternate start-order each round (L005); RNG result applied as noise (L006); fix misleading Force-id comment (CTR-006)
- process_simd_flip.rs: guard empty-Vec index with .get() (SF-01); guard division by zero (SF-02)

122 lib tests pass; 0 new warnings.

+++ crates/raw-pipeline/src/jxl_casadecoder.rs
@@ -1209,8 +1209,9 @@ pub fn decode_jxtc_region(
             compute_tile_copy_rects(&header, tx, ty, rx, ry, rw, rh, tw, th)
         {
             for row in 0..oh {
-                let src_row_off = ((src_y + row) * tw + src_x) as usize * bpp;
-                let dst_row_off = ((dst_y + row) * rw + dst_x) as usize * bpp;
+                // Cast to usize before multiplying to avoid u32 wrap on 32-bit (WASM).
+                let src_row_off = ((src_y + row) as usize * tw as usize + src_x as usize) * bpp;
+                let dst_row_off = ((dst_y + row) as usize * rw as usize + dst_x as usize) * bpp;
                 let row_bytes = (ow as usize) * bpp;
 
                 dest[dst_row_off..dst_row_off + row_bytes]
